% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infoshare.R
\name{download_infoshare}
\alias{download_infoshare}
\title{Download one or more data series from Infoshare}
\usage{
download_infoshare(
  selenium_connection,
  series_ids,
  target_directory,
  target_filename = NULL,
  browser_dl_directory = "~/Downloads"
)
}
\arguments{
\item{selenium_connection}{A connection to a Selenium server created by start_selenium()}

\item{series_ids}{A character vector of Stats NZ series IDs}

\item{target_directory}{The directory downloaded CSV files will be saved}

\item{target_filename}{Name of downloaded CSV file. If NULL (the default), a generic name will be used.}

\item{browser_dl_directory}{The directory where downloaded files by the browser are saved}
}
\value{
TRUE if data was successfully downloaded, otherwise FALSE
}
\description{
Download one or more data series from Infoshare
}
