% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infoshare.R
\name{start_selenium}
\alias{start_selenium}
\title{Start Selenium server connection}
\usage{
start_selenium(
  selenium_port = 4567L,
  selenium_browser = "chrome",
  selenium_verbose = FALSE,
  ...
)
}
\arguments{
\item{selenium_port}{The port number to use for the connection to Selenium server (default 4567L)}

\item{selenium_browser}{The browser for Selenium to use (default Google Chroms)}

\item{selenium_verbose}{Should Selenium operate in verbose mode? (default FALSE)}

\item{...}{Other parameters passed to RSelenium::rsDriver() to set up the Selenium connection}
}
\value{
A connection to the Selenium server
}
\description{
Start Selenium server connection
}
