% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infoshare.R
\name{stop_selenium}
\alias{stop_selenium}
\title{Stop Selenium server}
\usage{
stop_selenium(selenium_connection)
}
\arguments{
\item{selenium_connection}{A connection to a Selenium server created by start_selenium()}
}
\value{
Nothing
}
\description{
Stop Selenium server
}
